requirements-sqlite.txt
Django==4.2.7
python-decouple==3.8
Pillow>=10.0.0
django-crispy-forms==2.1
crispy-bootstrap5==2024.2
whitenoise==6.6.0
